Update index.html
I took out the get-app and support sections and changed the html in the about and features sections. I didn't make any major changes to the screens section because I don't know what screenshots we're going to be sticking in there. I wasn't sure whether to take out the Twitter/Google/FB links at the end because I didn't know if we' be replacing them with anything so I just left them.